Create Cargo workspace with a CLI project.
Added the root `Cargo.toml` file to configure the workspace and ran `cargo new cli`. I updated the `resolver` to version 2, as `cargo new` complained it didn't align with the 2021 edition.

I'm currently using Cargo 1.67.1, which I happened to already have installed. However I tried to `rustup` to the current latest but got weird "Undefined reference to `getauxval`" errors, so I'm just not going to worry about that right now.