Uses the `v8` crate to execute `BUILD.js` files and print a log.
I tried for several hours to create a stronger abstraction around this. I wanted a `JsExecutor` struct with a `.execute` method, but this doesn't appear to be possible due to the ownership semantics of the data. It does not appear easily possible to put these structs into a `JsExecutor` while retaining all their mutable references. The best I could do was to create a function which wraps script execution and pass it around where needed. I suspect the design I really want (a `JsExecutor::from` which creates the relevant data and stores it on the `JsExecutor` struct) is not possible with standard borrow semantics and I need to find an alternative design to solve this problem, but I'm not clear on what that design actually is.

I found https://github.com/patrickpissurno/rust_embedded_v8_example/tree/ui which demonstrates some common use cases (notably defining a `print` function) in v8 which was very helpful. `console.log` does not error and implies that it does something, but I could not find a way to hook into this mechanism (`isolate.add_message_listener` is only for uncaught errors), so the next best thing is to create a `print` function from scratch.

This is good enough for playing around, but I'd definitely like to refactor this in the future into a stronger abstraction.